Data on the COVID-19 pandemic for the US and individual states.

The dataset contains data on the COVID-19 pandemic for the US and individual states.

allstates

Format

A data frame with 12,082 rows and 43 variables:

positive

Total number of confirmed plus probable cases of COVID-19 reported by the state or territory

positiveIncrease

Cases (confirmed plus probable) calculated based on the previous day value.

probableCases

Total number of probable cases of COVID-19 as reported by the state or territory

positiveCasesViral

Total number of unique people with a positive PCR or other approved nucleic acid amplification test (NAAT)

negative

Total number of unique people with a completed PCR test that returns negative

negativeTestsViral

Total number of completed PCR tests (or specimens tested) that return negative as reported by the state or territory.

pending

Total number of viral tests that have not been completed as reported by the state or territory.

positiveTestsViral

Total number of completed PCR tests (or specimens tested) that return positive as reported by the state or territory.

totalTestsPeopleViral

Total number of unique people tested at least once via PCR testing, as reported by the state or territory.

totalTestsViral

Total number of PCR tests (or specimens tested) as reported by the state or territory.

totalTestEncountersViral

Total number of people tested per day via PCR testing as reported by the state or territory.

negativeTestsPeopleAntibody

The total number of unique people with completed antibody tests that return negative as reported by the state or territory.

negativeTestsAntibody

The total number of completed antibody tests that return negative as reported by the state or territory.

positiveTestsAntibody

Total number of completed antibody tests that return positive as reported by the state or territory.

totalTestsPeopleAntibody

The total number of unique people who have been tested at least once via antibody testing as reported by the state or territory.

totalTestsAntibody

Total number of completed antibody tests as reported by the state or territory.

positiveTestsPeopleAntigen

Total number of unique people with a completed antigen test that returned positive as reported by the state or territory.

positiveTestsAntigen

Total number of completed antigen tests that return positive as reported by the state or territory.

totalTestsPeopleAntigen

Total number of unique people who have been tested at least once via antigen testing, as reported by the state or territory.

totalTestsAntigen

Total number of completed antigen tests, as reported by the state or territory.

hospitalizedCumulative

Total number of individuals who have ever been hospitalized with COVID-19.

inIcuCumulative

Total number of individuals who have ever been hospitalized in the Intensive Care Unit with COVID-19.

onVentilatorCumulative

Total number of individuals who have ever been hospitalized under advanced ventilation with COVID-19.

hospitalizedCurrently

Individuals who are currently hospitalized with COVID-19.

inIcuCurrently

Individuals who are currently hospitalized in the Intensive Care Unit with COVID-19.

onVentilatorCurrently

Individuals who are currently hospitalized under advanced ventilation with COVID-19.

hospitalizedIncrease

Daily increase in hospitalizedCumulative, calculated from the previous day’s value.

death

Total fatalities with confirmed OR probable COVID-19 case diagnosis

deathConfirmed

Total fatalities with confirmed COVID-19 case diagnosis

deathProbable

Total fatalities with probable COVID-19 case diagnosis

deathIncrease

Daily increase in death, calculated from the previous day’s value.

recovered

Total number of people that are identified as recovered from COVID-19.

covidTrackingProjectPreferredTotalTestUnits

Label indicating which units The COVID Tracking Project prefers for each state’s “New tests” chart and on the state’s history page on source website.

dataQualityGrade

The COVID Tracking Project grade of the completeness of the data reporting by a state.

date

Date on which data was collected by The COVID Tracking Project.

lastUpdateEt

Date and time in Eastern time the state or territory last updated the data.

totalTestResultsIncrease

Daily increase in totalTestResults, calculated from the previous days value.

state

Two-letter abbreviation for the state or territory.

states

The number of states and territories included in the US dataset for each day.

totalTestResults

estimate of national testing performance.

Source

https://covidtracking.com/data/download